+++ internal/vcs/git/mocks.go
@@ -20,6 +20,7 @@ var Mocks, emptyMocks struct {
 	ReadDir          func(commit api.CommitID, name string, recurse bool) ([]os.FileInfo, error)
 	ResolveRevision  func(spec string, opt *ResolveRevisionOptions) (api.CommitID, error)
 	Stat             func(commit api.CommitID, name string) (os.FileInfo, error)
+	GetObject        func(objectName string) (OID, ObjectType, error)
 }
 
 // ResetMocks clears the mock functions set on Mocks (so that subsequent tests don't inadvertently