+++ schema/settings_stringdata.go
@@ -22,6 +22,12 @@ const SettingsSchemaJSON = `{
           "type": "boolean",
           "default": false,
           "!go": { "pointer": true }
+        },
+        "searchStats": {
+          "description": "Enables a new page that shows language statistics about the results for a search query.",
+          "type": "boolean",
+          "default": false,
+          "!go": { "pointer": true }
         }
       },
       "group": "Experimental"