+++ shared/src/e2e/driver.ts
@@ -11,6 +11,7 @@ import { readEnvBoolean, retry } from './e2e-test-utils'
 import * as path from 'path'
 import { escapeRegExp } from 'lodash'
 import { readFile } from 'mz/fs'
+import { Settings } from '../settings/settings'
 
 /**
  * Returns a Promise for the next emission of the given event on the given Puppeteer page.
@@ -413,18 +414,18 @@ export class Driver {
     }
 
     public async resetUserSettings(): Promise<void> {
+        return this.setUserSettings({})
+    }
+
+    public async setUserSettings<S extends Settings>(settings: S): Promise<void> {
         const currentSettingsResponse = await this.makeGraphQLRequest<IQuery>({
             request: gql`
                 query UserSettings {
                     currentUser {
                         id
-                        settingsCascade {
-                            subjects {
-                                latestSettings {
-                                    id
-                                    contents
-                                }
-                            }
+                        latestSettings {
+                            id
+                            contents
                         }
                     }
                 }
@@ -433,33 +434,29 @@ export class Driver {
         })
 
         const { currentUser } = dataOrThrowErrors(currentSettingsResponse)
+        if (!currentUser) {
+            throw new Error('no currentUser')
+        }
 
-        if (currentUser?.settingsCascade) {
-            const emptySettings = '{}'
-            const [{ latestSettings }] = currentUser.settingsCascade.subjects.slice(-1)
-
-            if (latestSettings && latestSettings.contents !== emptySettings) {
-                const updateConfigResponse = await this.makeGraphQLRequest<IMutation>({
-                    request: gql`
-                        mutation OverwriteSettings($subject: ID!, $lastID: Int, $contents: String!) {
-                            settingsMutation(input: { subject: $subject, lastID: $lastID }) {
-                                overwriteSettings(contents: $contents) {
-                                    empty {
-                                        alwaysNil
-                                    }
-                                }
+        const updateConfigResponse = await this.makeGraphQLRequest<IMutation>({
+            request: gql`
+                mutation OverwriteSettings($subject: ID!, $lastID: Int, $contents: String!) {
+                    settingsMutation(input: { subject: $subject, lastID: $lastID }) {
+                        overwriteSettings(contents: $contents) {
+                            empty {
+                                alwaysNil
                             }
                         }
-                    `,
-                    variables: {
-                        contents: emptySettings,
-                        subject: currentUser.id,
-                        lastID: latestSettings.id,
-                    },
-                })
-                dataOrThrowErrors(updateConfigResponse)
-            }
-        }
+                    }
+                }
+            `,
+            variables: {
+                contents: JSON.stringify(settings),
+                subject: currentUser.id,
+                lastID: currentUser.latestSettings ? currentUser.latestSettings.id : null,
+            },
+        })
+        dataOrThrowErrors(updateConfigResponse)
     }
 
     /**